moved insert record function from ban.py to models.py
